ai_engine/requirements.txt
torch
torchvision
torchaudio
diffusers
transformers
accelerate
sentencepiece
protobuf
safetensors
pillow
